Move lyric into content namespace.
